backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
stripe==7.12.0
alembic==1.13.1
asyncpg==0.29.0
